Updated use.rst because json.dumps raised error (#38)
The json.dumps line in "Using the Tool" section of Tatsu docs do not reflect the line in the README. An error was raised (raise TypeError(repr(o) + " is not JSON serializable")) when using json.dumps without asjson(). Updating that line to the code in README now works as intended.