[codegen] Use first define rule as staring rule, not 'start'
The documentation states that the starting rule is the first rule
defined in the grammar and the dynamically generated parser indeed
behaves like that. However, the generated code parser uses the hard
coded 'start' name for the start rule. Align the generated code to the
dynamic parser and to the documentation. Add test.