[model] Fall back to repr() for objects not handled by json module (#268)
The __str__() method of Node returns a JSON representation of the
Node. This fails if the Node contains anything that cannot be
represented by plain JSON. Notable examples are datetime and date
objects that are not so unrealistic for a parser to return.

Use a custom JSON encoder that fall back to repr() for these objects
so that a string representation for a Node can always be returned.

Fixes #262.