[walkers] Make implementation of NodeWalker more regular
NodeWalker defines a __new__() method that takes the unusual step of
modifying the class definition received as first argument adding a
_walker_cache class member. This means that all NodeWalker subclass
instances share the same _walker_cache and that this dictionary in
reset every time a new NodeWalker instance is created.

It may be that this implementation was an attempt to add instance
initialization without requiting subclasses to have to call
super().__init__() in their initialization. However, this results in
somehow bizarre behavior.

Changing the implementation to obtain a more regular and expected
behavior is worth the minuscule compatibility breakage of requiring
sublasses to call the parent __init__() method. Failing to do so
manifests with an obvious error and the fix is straightforward.

Fixes #266.