Add and state py38 compatibility (#288)
* RFC: Support for Python versions from 3.8 to 3.10 (#270)

* [model] Fall back to repr() for objects not handled by json module (#268)

The __str__() method of Node returns a JSON representation of the
Node. This fails if the Node contains anything that cannot be
represented by plain JSON. Notable examples are datetime and date
objects that are not so unrealistic for a parser to return.

Use a custom JSON encoder that fall back to repr() for these objects
so that a string representation for a Node can always be returned.

Fixes #262.

* [lint] fix linter warnings

* [codegen] Use modern super() call in generated ModelBuilderSemantics (#265)

Fixes #259.

* [docs] Small fixes and markup simplification (#264)

Fix a couple of markups errors, a typo, and a couple of broken links.

Simplify the markup using "::" code blocks instead than explicit
".. code::" blocks for TatSu syntax and disable syntax highlighting
for these.

To make markup errors similar to the ones fixed here less likely, turn
the description of TatSu expression into regular paragraphs rather
than quoted paragraphs.

* [docs] finx typo

* [test] fix failed tests

* [walkers] ensure one cache per walker class (#275)

fixes #266

* [codegen] make generated and dynamic parsers produce same AST

* [build] bump up version to avoid confusions

* [ast] fix pickling (@fizbin) (#276)

* [ast] fix pickling (@fizbin)

fixes #258

* [parsing] make the generated parser be the bootstrap parser

* [docs] minor bug fixing

* Release 5.8.1 (#277)

* [build] update version to 5.8.1
* [docs] update contributors
* [release] update CHANGELOG

* [docs] update copyright years

* [build/test] Run tests on Python versions from 3.8 to 3.10

* [build] Run CI on pushes on branches other than master

This allows to run the CI jobs on forks before submitting PRs.

* [setup] Relax required Python version to allow Python 3.8 and later

The tests do not evidence any dependence on Python syntax or features
introduced after Python 3.8. Having the tests run in the CI jobs on
all supported Python version will allow to easily catch any such
introduced dependence early so that the version check could be
tightened again.

* [tox] Extend Tox environments definition to include Python 3.8 and 3.9

* [build] Update version to 5.7.4

Co-authored-by: Juancarlo Añez <apalala@gmail.com>

* Update _version.py (#283)

Co-authored-by: Juancarlo Añez <apalala@gmail.com>

* [5.7.4] 5.8.2 equivalent release for Python 3.8 and later (#285)

Co-authored-by: Daniele Nicolodi <daniele@grinta.net>