Use temp directory instead of cwd for test storage (#370)
In `test_36_unichars` we are writing local python file and them import it for
testing purposes.

So far this was done in a local directory (current working dir), but this is
not ideal, as it might not be on a writable filesystem. For that reason it
seems better to use a temporary directory, that is much more commonly provided
in a writable form.

This change makes that shift.

Co-authored-by: Andrzej Fiedukowicz <afie@google.com>